+++ test/handler_test.dart
@@ -1,19 +1,21 @@
 import 'dart:async';
 
-import 'package:skiff/src/commands.dart';
+import 'package:skiff/src/commands/command.dart';
+import 'package:skiff/src/commands/command_handler.dart';
+import 'package:skiff/src/commands/command_result.dart';
 import 'package:test/test.dart';
 
-class StubCommand extends Command {
+class StubCommand implements Command {
   final String username;
   final String password;
 
-  StubCommand({this.username, this.password});
+  StubCommand(this.username, this.password);
 
   @override
-  String toString() => '${username}, ${"*" * password.length}';
+  String toString() => '$username, ${"*" * password.length}';
 }
 
-class StubCommandHandler implements CommandHandler<StubCommand> {
+class StubHandler implements CommandHandler<StubCommand> {
   bool _shouldFail = false;
 
   void fail() => _shouldFail = true;
@@ -27,31 +29,26 @@ class StubCommandHandler implements CommandHandler<StubCommand> {
 }
 
 void main() {
-  StubCommand stubCommand;
-  StubCommandHandler sut;
+  StubCommand? stubCommand;
+  StubHandler? sut;
 
   setUp(() {
-    stubCommand = StubCommand(username: 'jim.hop', password: 'h0pp3r');
-    sut = StubCommandHandler();
+    stubCommand = StubCommand('jim.hop', 'h0pp3r');
+    sut = StubHandler();
   });
 
-  group('StubCommandHandler', () {
+  group('StubHandler', () {
     test('.execute() returns successful result', () async {
-      // when
-      var value = await sut.execute(stubCommand);
+      var value = await sut!.execute(stubCommand!);
 
-      // then
       expect(value.isSuccessful, equals(true));
     });
 
     test('.execute() returns failed result', () async {
-      // given
-      sut.fail();
+      sut!.fail();
 
-      // when
-      var value = await sut.execute(stubCommand);
+      var value = await sut!.execute(stubCommand!);
 
-      // then
       expect(value.isSuccessful, equals(false));
     });
   });