+++ test/mediator_exception_test.dart
@@ -1,0 +1,12 @@
+import 'package:skiff/src/exceptions.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('MediatorException', () {
+    test('toString returns correct stringified exception message', () {
+      var exception = RequestTypeMissingException();
+      expect(exception.toString(),
+          'RequestTypeMissingException: Missing required request type');
+    });
+  });
+}