+++ test/mediator_test.dart
@@ -1,0 +1,99 @@
+import 'dart:async';
+
+import 'package:skiff/skiff.dart';
+import 'package:skiff/src/exceptions.dart';
+import 'package:skiff/src/handler.dart';
+import 'package:test/test.dart';
+
+import 'handler_test.dart';
+
+void main() {
+  FuncHandler<StubCommand, CommandResult>? stubCommandHandler;
+
+  Mediator? sut;
+
+  setUp(() {
+    stubCommandHandler = FuncHandler<StubCommand, CommandResult>(
+        (c) => Future.value(CommandResult.succeeded()));
+
+    sut = Mediator();
+  });
+
+  group('Mediator', () {
+    group('.addHandler', () {
+      test('throws exception if handler already exists', () {
+        sut!.addHandler<StubCommand>(stubCommandHandler!);
+
+        const expectedMessage = 'A handler already exists for StubCommand';
+
+        expect(
+            () => sut!.addHandler<StubCommand>(stubCommandHandler!),
+            throwsA(
+              isA<AlreadyRegisteredException>()
+                  .having((e) => e.message, 'message', equals(expectedMessage)),
+            ));
+      });
+
+      test('adds a handler for the given request type', () {
+        sut!.addHandler(stubCommandHandler!);
+        expect(sut!.handlers.length, 1);
+      });
+    });
+
+    group('.removeHandler', () {
+      test('throws exception if type missing', () {
+        const expectedMessage = 'Missing required request type';
+
+        expect(
+            () => sut!.removeHandler(),
+            throwsA(
+              isA<RequestTypeMissingException>()
+                  .having((e) => e.message, 'message', equals(expectedMessage)),
+            ));
+      });
+
+      test('throws exception if handler is missing', () {
+        const expectedMessage = 'Handler missing for StubCommand';
+
+        expect(
+            () => sut!.removeHandler<StubCommand>(),
+            throwsA(
+              isA<MissingHandlerException>()
+                  .having((e) => e.message, 'message', equals(expectedMessage)),
+            ));
+      });
+
+      test('removes handler for the given request type', () {
+        sut!.addHandler<StubCommand>(stubCommandHandler!);
+
+        var removedHandler = sut!.removeHandler<StubCommand>();
+
+        expect(removedHandler, equals(stubCommandHandler));
+        expect(sut!.handlers.length, 0);
+      });
+    });
+
+    group('.dispatch', () {
+      test('throws exception if handler is missing', () {
+        const expectedMessage = 'Handler missing for StubCommand';
+
+        var stubCommand = StubCommand('empty', 'empty');
+
+        expect(
+            () => sut!.dispatch<CommandResult>(stubCommand),
+            throwsA(
+              isA<MissingHandlerException>()
+                  .having((e) => e.message, 'message', equals(expectedMessage)),
+            ));
+      });
+
+      test('calls the handler and returns a response', () async {
+        sut!.addHandler<StubCommand>(stubCommandHandler!);
+
+        var response = await sut!.dispatch<CommandResult>(StubCommand('', ''));
+
+        expect(response.isSuccessful, isTrue);
+      });
+    });
+  });
+}